Added a few missing setcodes, Also.
It's probably time to make a real switch to the SCRYFALL api, and work with that.

It's going to be a sizeable overhaul and rework the card mechanic a ton.  I will need to set a timer such that we don't ping it until like 50-100ms has occurred between usages.

A fine change, but I hope scryfall is more effective than this one.  I think it will be honestly.
That's the next big plan.  Scryfall implementation and card reworking.