Scryfall is infinitely more effective than the alternative.
This has been incredibly easy to work with and not a problem at all.  I can overload the "response" object to handle everything appropriately as needed if I wanted to, I figure I'd be cleaner in the implementation and have separate structures to deal with everything

there will be so much code cleaning as time goes on but here we go.