Fixed Color typing issue and color -NOT issue.
Toughness,Power, and CMC all need the same logic, you will need a REGEX operation to use this without breaking the entire thing.
One single regex expression compiled, that we can reuse repeadetly to cycle and modify everything....

If you can make it generic (since we're just refining the issue further downward)
AKA, since we already get the right list of things from our regex^1 call, we can do this a level deeper and keep it generic for the most part.  Should be fine I believe, just have to ensure that we don't use the regexp for ALL characters (since that will break part of this)

Lets go.  Monday/Tuesday we will sort and separate this out into something that more closely resembles proper architecting.  But for now lets get the querying functionality up to speed.